requirements.txt
streamlit
pandas 
numpy
sentence-transformers
google.generativeai
pinecone-client
python-dotenv